+++ pandalone/xlsutils.py
@@ -41,7 +41,7 @@ def get_active_workbook():
 
 def get_Workbook(wrkb_fname):
     """
-    :param str wrkb_fname: if missing return the active excel-workbook 
+    :param str wrkb_fname: if missing return the active excel-workbook
     """
     import xlwings as xw
 
@@ -65,8 +65,8 @@ def _get_xl_vb_project(xl_wb):
       6. Click OK to close the Excel Options dialog box.
 
     or use the Excel-2010 shortcuts
-      [ALT] + [t][m][s] 
-    and then click the above check box. 
+      [ALT] + [t][m][s]
+    and then click the above check box.
     """
     from win32com.universal import com_error
     import easygui
@@ -207,7 +207,7 @@ def import_files_into_excel_workbook(infiles_wildcard, wrkb_fname=None, new_fnam
 
     :param str infiles_wildcard: 'some/foo*bar.vba'
     :param str wrkb_fname: filepath to update, active-workbook assumed if missing
-    :param str new_fname: a macro-enabled (`.xlsm`, `.xltm`) filepath for updated workbook, 
+    :param str new_fname: a macro-enabled (`.xlsm`, `.xltm`) filepath for updated workbook,
                             `.xlsm` gets appended if current not a macro-enabled excel-format
     """
 
@@ -239,150 +239,6 @@ def normalize_local_url(self, urlparts):
     return urlparts.geturl()
 
 
-#############################
-# Excel-refs
-#############################
-
-
-# TODO Transform resolve_excel_ref() code into xlwings xlasso-backend.
-def resolve_excel_ref(ref_str, default=None, _cntxtx=None):
-    """
-    Parses and fetches the contents of an `excel_ref` (the hash-part of the `excel_url`).
-    if `ref_str` is an *excel-ref*, it returns the referred-contents as DataFrame or a scalar, `None` otherwise.
-
-    Excel-ref examples::
-
-        #a1
-        #E5.column
-        #some sheet_name!R1C5.TABLE
-        #1!a1:c5.table(header=False)
-        #3!a1:C5.horizontal(strict=True; atleast_2d=True)
-        #sheet-1!A1.table(asarray=True){columns=['a','b']}
-        #any%sheet^&name!A1:A6.vertical(header=True)        ## Setting Range's `header` kw and 
-                                                            #      DataFrame will parse 1st row as header
-
-    The *excel-ref* syntax is case-insensitive apart from the key-value pairs, 
-    which it is given in BNF-notation:
-
-    .. productionlist::
-            excel_ref   : "#" 
-                        : [sheet "!"] 
-                        : cells 
-                        : ["." shape] 
-                        : ["(" range_kws ")"] 
-                        : ["{" df_kws "}"]
-            sheet       : sheet_name | sheet_index
-            sheet_name  : <any character>
-            sheet_index : `integer`
-            cells       : cell_ref [":" cell_ref]
-            cell_ref    : A1_ref | RC_ref | tuple_ref
-            A1_ref      : <ie: "A1" or "BY34">
-            RC_ref      : <ie: "R1C1" or "R24C34">
-            tuple_ref   : <ie: "(1,1)" or "(24,1)", the 1st is the row>
-            shape       : "." ("table" | "vertical" | "horizontal")
-            range_kws   : kv_pairs                    # keywords for xlwings.Range(**kws)
-            df_kws      : kv_pairs                    # keywords for pandas.DataFrafe(**kws)
-            kv_pairs    : <python code for **keywords ie: "a=3.14, f = 'gg'">
-
-
-    Note that the "RC-notation" is not converted, so Excel may not support it (unless overridden in its options).
-    """
-    import xlwings as xw
-
-    matcher = _excel_ref_specifier_regex.match(ref_str)
-    if matcher:
-        ref = matcher.groupdict()
-        log.info("Parsed string(%s) as Excel-ref: %s", ref_str, ref)
-
-        sheet = ref.get('sheet') or ''
-        try:
-            sheet = int(sheet)
-        except ValueError:
-            pass
-        cells = ref['ref']
-        range_kws = _parse_kws(ref.get('range_kws'))
-        ref_range = xw.Range(sheet, cells, **range_kws)
-        range_shape = ref.get('shape')
-        if range_shape:
-            ref_range = operator.attrgetter(range_shape.lower())(ref_range)
-
-        v = ref_range.value
-
-        if ref_range.row1 != ref_range.row2 or ref_range.col1 != ref_range.col2:
-            # Parse as DataFrame when more than one cell.
-            #
-            pandas_kws = _parse_kws(ref.get('pandas_kws'))
-            if 'header' in range_kws and not 'columns' in pandas_kws:
-                # Do not ignore explicit-columns.
-                v = pd.DataFrame(v[1:], columns=v[0], **pandas_kws)
-            else:
-                v = pd.DataFrame(v, **pandas_kws)
-
-        log.debug("Excel-ref(%s) value: %s", ref, v)
-
-        return v
-    else:
-        if default is _undefined:
-            raise ValueError("Invalid excel-ref(%s)!" % ref_str)
-        else:
-            return default
-
-
-def test_excel_refs(self):
-    from pandalone.xlsutils import resolve_excel_ref
-    sheetname = 'Input'
-    addr = 'd2'
-    table = pd.DataFrame(
-        {'a': [1, 2, 3], 'b': ['s', 't', 'u'], 'c': [True, False, True]})
-    table.index.name = 'id'
-
-    cases = [
-        ("@d2",                         'id'),
-        ("@e2",                         'a'),
-        ("@e3",                         1),
-        ("@e3:g3",
-         table.iloc[0, :].values.reshape((3, 1))),
-        ("@1!e2.horizontal",
-         table.columns.values.reshape((3, 1))),
-        ("@1!d3.vertical",
-         table.index.values.reshape((3, 1))),
-        ("@e4.horizontal",
-         table.iloc[1, :].values.reshape((3, 1))),
-        ("@1!e3:g3.vertical",           table.values),
-        ("@{sheet}!E2:F2.table(strict=True, header=True)".format(sheet=sheetname),
-         table),
-    ]
-
-    errors = []
-    with xw_Workbook() as wb:
-        _make_sample_sheet(wb, sheetname, addr, table)
-        for i, (inp, exp) in enumerate(cases):
-            out = None
-            try:
-                out = resolve_excel_ref(inp)
-                log.debug(
-                    '%i: INP(%s), OUT(%s), EXP(%s)', i, inp, out, exp)
-                if exp is not None:
-                    if isinstance(out, NDFrame):
-                        out = out.values
-                    if isinstance(exp, NDFrame):
-                        exp = exp.values
-                    if isinstance(out, np.ndarray) or isinstance(exp, np.ndarray):
-                        npt.assert_array_equal(
-                            out, exp, '%i: INP(%s), OUT(%s), EXP(%s)' % (i, inp, out, exp))
-                    else:
-                        self.assertEqual(
-                            out, exp, '%i: INP(%s), OUT(%s), EXP(%s)' % (i, inp, out, exp))
-            except Exception as ex:
-                log.exception(
-                    '%i: INP(%s), OUT(%s), EXP(%s), FAIL: %s' % (i, inp, out, exp, ex))
-                errors.append(ex)
-
-    if errors:
-        raise Exception('There are %i out of %i errors!' %
-                        (len(errors), len(cases)))
-
-
 def main(*argv):
     """
     Updates the vba code on excel workbooks.
@@ -392,8 +248,8 @@ def main(*argv):
         {cmd}  *.vbas                       # Add modules into active excel-workbook.
         {cmd}  *.vbas  in_file.xlsm         # Specify input workbook to add/update.
         {cmd}  *.vbas  foo.xlsx  bar.xlsm   # Specify input & output workbooks.
-        {cmd}  *.vbas  foo.xls              # A `foo.xls.xlsm` will be created. 
-        {cmd}  *.vbas  foo.xls   bar        # A `bar.xlm` will be created. 
+        {cmd}  *.vbas  foo.xls              # A `foo.xls.xlsm` will be created.
+        {cmd}  *.vbas  foo.xls   bar        # A `bar.xlm` will be created.
         {cmd}  --pandalone inp.xls [out]    # Add 'pandalone-xlwings' modules.
     """